+++ app/src/main/java/org/CashCatalysts/CashCatalysts/Transactions/TransactionHandler.java
@@ -1,5 +1,6 @@
 package org.CashCatalysts.CashCatalysts.Transactions;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
@@ -8,7 +9,7 @@ public class TransactionHandler {
     private List<Transaction> transactions = new ArrayList<>();
     private int currentId = 1; // To generate unique transaction IDs
 
-    public void addTransaction(String expenseOrIncome, String name, String type, Date date, double amount) {
+    public void addTransaction(String expenseOrIncome, String name, String type, Date date, BigDecimal amount) {
         Transaction transaction = new Transaction(currentId++, expenseOrIncome, name, type, date, amount);
         transactions.add(transaction);
     }
